+++ core/components/svgsanitizer/docs/changelog.md
@@ -5,6 +5,7 @@ Released on ...
 
 - Require PHP 8
 - Update svg-sanitize to 0.20.0
+- Add option to change the cacheKey
 - Add width and height to SVG based on viewBox
 
 ## svgSanitizer 1.1.0